promises
promiseone and promisetwo: These promises resolve after a delay and demonstrate the use of .then() and .catch() for handling asynchronous tasks.
promisethree: This promise resolves an object after 97ms, and the resolved value is logged.
promisefour: Introduces a simulated error handling mechanism, where the promise either resolves with an object of capitals or rejects with an error message. It also demonstrates chaining .then() calls.
promisefive: Uses async/await to simplify the handling of promises. The asyncawait function waits for the promise to resolve and catches errors.
GitHub API Fetch: Fetches data from a GitHub user API and logs the response or any errors encountered.