inheritance
Key Points Explained:
Class Definition:

The User class is defined with a constructor that initializes a username property. It has a method naming() that returns a formatted string containing the username.
Inheritance:

The Company class extends the User class, inheriting its properties and methods. The Company constructor calls the parent constructor using super() and initializes additional properties like email and password.
Method Overriding:

The Company class has its own method addCourse(), which utilizes the inherited username from the User class to return a string indicating who added the course.
Instances:

Two instances are created: asd of type Company and gulmol of type User. The methods naming() and addCourse() are called on these instances to demonstrate their functionality.
Minor Improvements:
Naming Conventions: Class names in JavaScript are conventionally written in PascalCase. I changed user to User and company to Company to adhere to this convention.
Method Naming: The method addcourse was renamed to addCourse to maintain camelCase style for method names.