mathpi
Key Points Explained:
Property Descriptor:

The Object.getOwnPropertyDescriptor method returns a property descriptor object for a specific property. It provides details such as whether the property is writable, enumerable, and configurable.
Attempting to Modify a Built-in Property:

You attempted to change the value of Math.PI. This operation does not succeed because Math.PI is a read-only property in JavaScript, meaning it cannot be modified.
Object Creation:

A rice object is created with properties (name, price, majorProducer) and a method (orderRice). Note the fix to the property name to follow camelCase conventions.
Defining Property Descriptors:

Object.defineProperty is used to change the attributes of the name property. After calling it, the name property can be modified (writable: true), will not show up when iterating through the object (enumerable: false), and cannot be deleted or redefined (configurable: false).
Iterating Over Object Properties:

The code uses a for...of loop with Object.entries to iterate over the properties of the rice object. It only logs properties that are not functions, thus excluding the orderRice method from the output.
Output Example
The Object.getOwnPropertyDescriptor logs will show details about the property configurations.
The loop will print:

name: basmati rice
price: 300
majorProducer: china, usa