object
Explanation of Key Concepts
Function Definition:

mul(num): A simple function that multiplies the input number by 8.
Prototype:

creatuser: A constructor function that creates user objects with properties username and score.
Prototypes allow you to define methods that can be shared among all instances of the object created by the constructor. The increment method increments the user's score, and the printme method logs the user's score.
Creating Instances:

When you create an instance of creatuser using the new keyword, a new object is created, the creatuser constructor is called with this bound to the new object, and the prototype methods are available on that object.
Using Methods:

After creating the kelvin and kelvins instances, you call their respective increment and printme methods to update and display their scores.