prototype
Explanation of Key Concepts
Constructor Function:

The Country function is a constructor that creates country objects with specific properties and methods.
Prototype Inheritance:

The __proto__ property is used to establish a prototype chain, allowing properties and methods from one object to be accessible in another object.
Object and Array Prototypes:

New methods (fin for objects and fini for arrays) are added to the prototypes, which means all objects and arrays can access these methods.
Logging Prototype Chain:

The console.log statements demonstrate how properties can be accessed from different levels of the prototype chain.
String Prototype Method:

The trueLength method is added to the String prototype to calculate the length of the string after trimming whitespace. This shows how we can extend built-in types in JavaScript.
Output
When you run this code, the output will reflect the statements in the console.log functions, indicating property values from the objects and demonstrating the prototype chain's functionality. If you uncomment the line with euCountryNames.fin(), it will throw an error because that method is not directly available to the euCountryNames array.