staticprop.
Explanation of Key Concepts
Class Definition:

The Animal class has a constructor that initializes the name property and a method (logName) to log the name of the animal.
Static Method:

The createSpeciesId method is static, meaning it belongs to the class itself and not to instances of the class. Therefore, it can only be called on the class (e.g., Cat.createSpeciesId()) and not on an instance (e.g., persianCat.createSpeciesId()).
Inheritance:

The Cat class extends the Animal class, inheriting its properties and methods. It also has its own property (breed) and method (logBreed) to log the breed.
Creating Instances:

Instances of Animal and Cat are created using the new keyword. Methods defined in the class can be called on these instances.
Output
When you run this code, the output will be:

Animal-001
Animal Name: Persian
Breed: Persian Cat

Notes
The commented-out lines that try to call createSpeciesId() on instances will throw an error because static methods should be accessed via the class name.
The code clearly demonstrates how static methods and inheritance work in JavaScript classes. If you have any further questions or would like to explore additional concepts, feel free to ask!