Updated code with comment in html
Wooden Project

This project features a simple web page designed to provide a welcoming experience with buttons for starting and stopping an action. It utilizes HTML, CSS, and JavaScript to create an interactive user interface.

File Descriptions:
HTML (index.html):

This file contains the structure of the web page. It features a heading welcoming the user and a container for buttons.
The "Start" and "Stop" buttons will trigger actions defined in the linked JavaScript file.
CSS Styles:

The internal CSS styles create a clean, centered layout with a light lavender background and pink button designs.
The flexbox layout is used for both the body and the button container to ensure proper alignment and spacing.
Usage
To use this project, open the index.html file in a web browser. Clicking the "Start" and "Stop" buttons will trigger actions defined in the color.js JavaScript file, which you can implement to add interactivity to the buttons.