Updated code with comment in js
Wooden Project - Color Changing Background

This JavaScript file contains functions to generate random colors and change the background color of the webpage when the "Start" button is clicked. The background color will change every 1717 milliseconds until the "Stop" button is pressed, which will revert the background color to its original state.

File Descriptions:
JavaScript (color.js):
randomcolor: Generates a random hex color code by concatenating six random hex digits.
changebackgroundcolor: Starts an interval that changes the background color of the webpage to a random color every 1717 milliseconds, only if the background color is not already changing.
stopbackgroundcolor: Stops the interval that changes the background color and resets it to the original color (#E6E6FA).
Event listeners are added to the "Start" and "Stop" buttons to control the background color changes.
Usage
To use this functionality, ensure the color.js file is linked in your HTML file. When you click the "Start" button, the background color of the webpage will start changing randomly. Clicking the "Stop" button will halt the changes and reset the background to its original color.